refactor: remove the FixBitSet depenedency (#59)
We use a u16 for players in a table. So this was the only place still remaining that was using fixed bit set. It didnt' seem worth it to keep for one place.

This ends up being a rather large perf win for ICM since we don't generate any iterators or 2 element vectors from choose random.