perf: remove uuid (#182)
* perf: remove uuid

Summary:
Since I was looking at the flamegraph svg. I noticed a good ammount of
uuid time on cpu. We don't need that complexity. We only need an ID's
not going to repeat with no-coordination. This adds that by using u128

** This is a breaking change. **

It changes the Agent API
It changes the Historian API

Test:
hyperfine tests

* chore: fmt